fix: changes after feedback:
* moved context into separate file
* add syncPlayheadInfinitesForNextPartInstance
* proper use arguments to postProcessPieces
