+++ meteor/server/api/blueprints/context/index.ts
@@ -1,2 +1,3 @@
 export * from './context'
 export * from './adlibActions'
+export * from './syncIngestUpdateToPartInstance'
